third_party/WebKit/LayoutTests/external/wpt/cssom-view/cssom-getBoundingClientRect-002-expected.txt

This is a testharness.js-based test.
FAIL getBoundingClientRect on a newly-created Element not yet inserted into the DOM should return an all-zeroes DOMRect assert_equals: DOMRect's x should be zero expected (number) 0 but got (undefined) undefined
Harness: the test ran to completion.
